+++ src/lib/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./copyTextToClipboard";
 export * from "./formatDate";
-export * from "./device-uuid";
